19.23 15.03.2022

+++ src/main/java/by/suprun/task2/entity/AbstractTariff.java
@@ -72,7 +72,7 @@ public abstract class AbstractTariff {
         this.monthPayRoll = monthPayRoll;
     }
 
-    public long getSmsPrise() {
+    public int getSmsPrise() {
         return smsPrise;
     }
